+++ fili-core/src/main/java/com/yahoo/bard/webservice/table/ConcretePhysicalTable.java
@@ -4,9 +4,9 @@ package com.yahoo.bard.webservice.table;
 
 import com.yahoo.bard.webservice.data.config.names.TableName;
 import com.yahoo.bard.webservice.data.time.ZonedTimeGrain;
-import com.yahoo.bard.webservice.table.availability.ImmutableAvailability;
+import com.yahoo.bard.webservice.metadata.DataSourceMetadataService;
+import com.yahoo.bard.webservice.table.availability.ConcreteAvailability;
 
-import java.util.Collections;
 import java.util.Map;
 
 import javax.validation.constraints.NotNull;
@@ -22,19 +22,21 @@ public class ConcretePhysicalTable extends BasePhysicalTable {
      * @param timeGrain time grain of the table
      * @param columns the columns for this table
      * @param logicalToPhysicalColumnNames mappings from logical to physical names
+     * @param metadataService datasource metadata service containing availability data for the table
      */
     public ConcretePhysicalTable(
             @NotNull TableName name,
             @NotNull ZonedTimeGrain timeGrain,
             @NotNull Iterable<Column> columns,
-            @NotNull Map<String, String> logicalToPhysicalColumnNames
+            @NotNull Map<String, String> logicalToPhysicalColumnNames,
+            @NotNull DataSourceMetadataService metadataService
     ) {
         super(
                 name,
                 timeGrain,
                 columns,
                 logicalToPhysicalColumnNames,
-                new ImmutableAvailability(name, Collections.emptyMap())
+                new ConcreteAvailability(name, metadataService)
         );
     }
 
@@ -44,7 +46,8 @@ public class ConcretePhysicalTable extends BasePhysicalTable {
      * @param name name of the physical table as String
      * @param timeGrain time grain of the table
      * @param columns the columns for this table
-     * @param logicalToPhysicalColumnNames  mappings from logical to physical names
+     * @param logicalToPhysicalColumnNames mappings from logical to physical names
+     * @param metadataService datasource metadata service containing availability data for the table
      *
      * @deprecated Should use constructor with TableName instead of String as table name
      */
@@ -53,9 +56,10 @@ public class ConcretePhysicalTable extends BasePhysicalTable {
             @NotNull String name,
             @NotNull ZonedTimeGrain timeGrain,
             @NotNull Iterable<Column> columns,
-            @NotNull Map<String, String> logicalToPhysicalColumnNames
+            @NotNull Map<String, String> logicalToPhysicalColumnNames,
+            @NotNull DataSourceMetadataService metadataService
     ) {
-        this(TableName.of(name), timeGrain, columns, logicalToPhysicalColumnNames);
+        this(TableName.of(name), timeGrain, columns, logicalToPhysicalColumnNames, metadataService);
     }
 
     public String getFactTableName() {