+++ fili-core/src/main/java/com/yahoo/bard/webservice/table/PhysicalTable.java
@@ -6,11 +6,12 @@ import com.yahoo.bard.webservice.data.config.names.TableName;
 import com.yahoo.bard.webservice.data.time.ZonedTimeGrain;
 import com.yahoo.bard.webservice.druid.model.query.Granularity;
 import com.yahoo.bard.webservice.table.availability.Availability;
+import com.yahoo.bard.webservice.table.resolver.DataSourceConstraint;
+import com.yahoo.bard.webservice.util.SimplifiedIntervalList;
 
 import org.joda.time.DateTime;
 import org.joda.time.Interval;
 
-import java.util.List;
 import java.util.Map;
 import java.util.Set;
 
@@ -29,41 +30,38 @@ public interface PhysicalTable extends Table {
     /**
      * Get the name of the current table.
      *
-     * @return name of the table as TableName
+     * @return name of the table as a TableName
      */
     TableName getTableName();
 
     /**
      * Get the value of the actual availability for this physical table.
      *
-     * @return The current actual physical availability or a runtime exception if there isn't one yet.
+     * @return the current actual physical availability or a runtime exception if there isn't one yet.
      */
     Availability getAvailability();
 
     /**
+     * Get the granularity for the table.
+     *
+     * @return the granularity of this table
+     */
+    Granularity getGranularity();
+
+    /**
      * Get a date time that the table will align to based on grain and available intervals.
      *
-     * @return The time of either the first available interval of any columns in this table or now, floored to the
+     * @return the time of either the first available interval of any columns in this table or now, floored to the
      * table's time grain.
      */
     DateTime getTableAlignment();
 
     /**
-     * Determine whether or not this PhysicalTable has a mapping for a specific logical name.
+     * Getter for all possible column available intervals.
      *
-     * @param logicalName  Logical name to check
-     * @return True if contains a non-default mapping for the logical name, false otherwise
+     * @return tableEntries map of column to set of available intervals
      */
-    boolean hasLogicalMapping(String logicalName);
-
-    /**
-     * Fetch a set of intervals given a column name.
-     *
-     * @param columnName  Name of the column
-     *
-     * @return Set of intervals associated with a column, empty if column is missing
-     */
-    Set<Interval> getIntervalsByColumnName(String columnName);
+    Map<Column, Set<Interval>> getAllAvailableIntervals();
 
     /**
      * Translate a logical name into a physical column name. If no translation exists (i.e. they are the same),
@@ -76,43 +74,38 @@ public interface PhysicalTable extends Table {
      * such as dimensions used for DimensionSpecs in queries to do mapping from fact-level dimensions to something else,
      * should likely use their own serialization strategy so as to not hit this defaulting behavior.
      *
-     * @param logicalName  Logical name to lookup in physical table
-     * @return Translated logicalName if applicable
+     * @param logicalName logical name to lookup in physical table
+     *
+     * @return translated logicalName if applicable
      */
     String getPhysicalColumnName(String logicalName);
 
     /**
+     * Getter for active column intervals.
+     *
+     * @param constraints data constraint containing columns and api filters
+     *
+     * @return tableEntries map of column to set of available intervals
+     */
+    SimplifiedIntervalList getAvailableIntervals(DataSourceConstraint constraints);
+
+    /**
      * Get the columns from the schema for this physical table.
      *
-     * @return The columns of this physical table
+     * @return the columns of this physical table
      *
-     * @deprecated In favor of getting the columns directly from the schema
+     * @deprecated in favor of getting the columns directly from the schema
      */
     @Deprecated
     Set<Column> getColumns();
 
     /**
-     * Getter for active column intervals.
-     *
-     * @return tableEntries map of column to set of available intervals
-     *
-     */
-    Map<Column, List<Interval>> getAvailableIntervals();
-
-    /**
      * Get the time grain from granularity.
      *
-     * @return The time grain of this physical table
+     * @return the time grain of this physical table
      *
      * @deprecated use getSchema().getGranularity()
      */
     @Deprecated
     ZonedTimeGrain getTimeGrain();
-
-    /**
-     * Get the granularity for the table.
-     *
-     * @return The granularity of this table
-     */
-    Granularity getGranularity();
 }