+++ fili-core/src/main/java/com/yahoo/bard/webservice/table/availability/Availability.java
@@ -3,12 +3,11 @@
 package com.yahoo.bard.webservice.table.availability;
 
 import com.yahoo.bard.webservice.data.config.names.TableName;
-import com.yahoo.bard.webservice.table.Column;
+import com.yahoo.bard.webservice.table.resolver.DataSourceConstraint;
+import com.yahoo.bard.webservice.util.SimplifiedIntervalList;
 
 import org.joda.time.Interval;
 
-import java.util.Collections;
-import java.util.List;
 import java.util.Map;
 import java.util.Set;
 
@@ -25,30 +24,18 @@ public interface Availability {
     Set<TableName> getDataSourceNames();
 
     /**
-     * The availability for a given column.
-     *
-     * @param c  A column
-     *
-     * @return The list of intervals that column is available for.
-     */
-    List<Interval> get(Column c);
-
-    /**
      * The availability of all columns.
      *
-     * @return The intervals, by column, available.
+     * @return the intervals, by column, available.
      */
-    Map<Column, List<Interval>> getAvailableIntervals();
+    Map<String, Set<Interval>> getAllAvailableIntervals();
 
     /**
-     * Fetch a set of intervals given a column name.
+     * Fetch a set of intervals given a set of column name.
      *
-     * @param columnName  Name of the column
+     * @param constraints data constraint containing columns and api filters
      *
-     * @return Set of intervals associated with a column, empty if column is missing
+     * @return set of intervals associated with each corresponding column in a map, empty if column is missing
      */
-    default List<Interval> getIntervalsByColumnName(String columnName) {
-        List<Interval> result = get(new Column(columnName));
-        return result == null ? Collections.emptyList() : result;
-    }
+    SimplifiedIntervalList getAvailableIntervals(DataSourceConstraint constraints);
 }