+++ fili-core/src/test/groovy/com/yahoo/bard/webservice/druid/model/query/TopNQuerySpec.groovy
@@ -13,6 +13,7 @@ import com.yahoo.bard.webservice.data.dimension.impl.ScanSearchProviderManager
 import com.yahoo.bard.webservice.druid.model.aggregation.Aggregation
 import com.yahoo.bard.webservice.druid.model.datasource.TableDataSource
 import com.yahoo.bard.webservice.druid.model.postaggregation.PostAggregation
+import com.yahoo.bard.webservice.metadata.DataSourceMetadataService
 import com.yahoo.bard.webservice.table.ConcretePhysicalTable
 import com.yahoo.bard.webservice.util.GroovyTestUtils
 
@@ -46,9 +47,9 @@ class TopNQuerySpec extends Specification {
         vars.dataSource = vars.dataSource ?: new TableDataSource<TopNQuery>(new ConcretePhysicalTable(
                 "table_name",
                 DAY.buildZonedTimeGrain(DateTimeZone.UTC),
-                [] as Set
-                ,
-                ["apiLocale": "locale"]
+                [] as Set,
+                ["apiLocale": "locale"],
+                Mock(DataSourceMetadataService)
         ))
         vars.dimension = vars.dimension ?: ""
         vars.threshold = vars.threshold ?: 5